Remove debian/patches/upstream/0010-GStreamer-support-video-orientation.patch and reorder patches
It's about getting orientation data from video, which is not related to
camera. qtubuntu-media automagically rotate the video as it gets
orientation data, which is the same behavior as the new nemo-
qtmultimedia-plugins' video backend. That is, developers is never
expected to handle orientation themself. So, enabling this will add
confusions to developers.